Rename src files
