+++ python/fibonacci_sequence/src/application.py
